moves lightwallet into sperate folder
