+++ app/components/walletSelector/walletLight/WalletLight.tsx
@@ -1,0 +1,10 @@
+import * as React from "react";
+import { Container } from "reactstrap";
+
+import { WalletLightRouter } from "./WalletLightRouter";
+
+export const WalletLight: React.SFC<{}> = () => (
+  <Container>
+    <WalletLightRouter />
+  </Container>
+);
