+++ app/components/walletSelector/walletLight/walletLightRoutes.tsx
@@ -1,9 +1,9 @@
-import { walletRoutes } from "./walletRoutes";
+import { walletRoutes } from "../walletRoutes";
 
 const parentRoutePath = walletRoutes.light;
 
 export const walletLightRoutes = {
   create: parentRoutePath + "/create",
   recover: parentRoutePath + "/recover",
-  save: parentRoutePath + "/save",
+  validate: parentRoutePath + "/validate",
 };
